ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.3.1
spacy==3.7.2
langdetect==1.0.9
python-multipart==0.0.6
redis==5.0.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.1.0
Pillow==10.2.0
pytesseract==0.3.10
numpy==1.26.3
scikit-learn==1.4.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
prometheus-client==0.19.0
structlog==24.1.0
